+++ src/relax/op/nn/nn.h
@@ -65,7 +65,7 @@ Expr batch_norm(Expr data, Expr gamma, Expr beta, Expr moving_mean, Expr moving_
                 int axis, double epsilon, bool center, bool scale);
 
 /*! \brief Compute local response normalization (LRN) */
-Expr lrn(Expr data, int size, int axis, float alpha, float beta, float bias);
+Expr lrn(Expr data, int size, int axis, double alpha, double beta, double bias);
 
 /*! \brief Compute layer normalization. */
 Expr layer_norm(Expr data, Expr gamma, Expr beta, Array<Integer> axes, double epsilon, bool center,
